refactor: remove dead code and consolidate duplicates
- Remove unused XInput keystroke API (~68 lines)
- Remove unused FromGame() method from OverlayItem
- Remove unused GetWindowTextLength P/Invoke
- Remove unused mouse message constants from InputBlocker
- Remove unused TertiaryText property from OverlayItem
- Remove legacy OverlayUI project (empty scaffold)
- Create ProcessMatchingUtils for shared utilities
- Consolidate duplicate LauncherProcessNames, GetGameNameWords,
  IsProcessNameMatch, IsWindowTitleMatch, GetBestImagePath
- Update CONTRIBUTING.md to reflect new structure
