+++ src/OverlayPlugin/Interop/XInput.cs
@@ -170,72 +170,4 @@ internal static class XInput
         return TryGetState(userIndex, out state);
     }
 
-    // ========== Keystroke APIs ==========
-    // Note: XInputGetKeystroke does NOT report Guide button events.
-    // Use TryGetStateEx and check XINPUT_GAMEPAD_GUIDE bit instead.
-
-    [StructLayout(LayoutKind.Sequential)]
-    public struct XINPUT_KEYSTROKE
-    {
-        public ushort VirtualKey;
-        public char Unicode;
-        public ushort Flags;
-        public byte UserIndex;
-        public byte HidCode;
-    }
-
-    public const ushort XINPUT_KEYSTROKE_KEYDOWN = 0x0001;
-    public const ushort XINPUT_KEYSTROKE_KEYUP = 0x0002;
-    public const ushort XINPUT_KEYSTROKE_REPEAT = 0x0004;
-
-    // VK_PAD constants for XInputGetKeystroke (NOT for Guide button!)
-    public const ushort VK_PAD_A = 0x5800;
-    public const ushort VK_PAD_B = 0x5801;
-    public const ushort VK_PAD_X = 0x5802;
-    public const ushort VK_PAD_Y = 0x5803;
-    public const ushort VK_PAD_RSHOULDER = 0x5804;
-    public const ushort VK_PAD_LSHOULDER = 0x5805;
-    public const ushort VK_PAD_LTRIGGER = 0x5806;
-    public const ushort VK_PAD_RTRIGGER = 0x5807;
-    public const ushort VK_PAD_DPAD_UP = 0x5810;
-    public const ushort VK_PAD_DPAD_DOWN = 0x5811;
-    public const ushort VK_PAD_DPAD_LEFT = 0x5812;
-    public const ushort VK_PAD_DPAD_RIGHT = 0x5813;
-    public const ushort VK_PAD_START = 0x5814;
-    public const ushort VK_PAD_BACK = 0x5815;
-    public const ushort VK_PAD_LTHUMB_PRESS = 0x5816;
-    public const ushort VK_PAD_RTHUMB_PRESS = 0x5817;
-
-    [DllImport("xinput1_4.dll", EntryPoint = "XInputGetKeystroke")]
-    private static extern int XInputGetKeystroke1_4(int dwUserIndex, int dwReserved, out XINPUT_KEYSTROKE pKeystroke);
-
-    [DllImport("xinput1_3.dll", EntryPoint = "XInputGetKeystroke")]
-    private static extern int XInputGetKeystroke1_3(int dwUserIndex, int dwReserved, out XINPUT_KEYSTROKE pKeystroke);
-
-    [DllImport("xinput9_1_0.dll", EntryPoint = "XInputGetKeystroke")]
-    private static extern int XInputGetKeystroke9_1_0(int dwUserIndex, int dwReserved, out XINPUT_KEYSTROKE pKeystroke);
-
-    public static bool TryGetKeystroke(int userIndex, out XINPUT_KEYSTROKE stroke)
-    {
-        try
-        {
-            if (XInputGetKeystroke1_4(userIndex, 0, out stroke) == ERROR_SUCCESS)
-                return true;
-        }
-        catch { }
-        try
-        {
-            if (XInputGetKeystroke1_3(userIndex, 0, out stroke) == ERROR_SUCCESS)
-                return true;
-        }
-        catch { }
-        try
-        {
-            if (XInputGetKeystroke9_1_0(userIndex, 0, out stroke) == ERROR_SUCCESS)
-                return true;
-        }
-        catch { }
-        stroke = default;
-        return false;
-    }
 }
