--- src/OverlayUI/App.xaml
@@ -1,8 +1,0 @@
-<Application x:Class="OverlayUI.App"
-             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             StartupUri="MainWindow.xaml">
-    <Application.Resources>
-    </Application.Resources>
-</Application>
-
