--- src/OverlayUI/App.xaml.cs
@@ -1,8 +1,0 @@
-using System.Windows;
-
-namespace OverlayUI;
-
-public partial class App : Application
-{
-}
-
