--- src/OverlayUI/MainWindow.xaml
@@ -1,26 +1,0 @@
-<Window x:Class="OverlayUI.MainWindow"
-        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-        WindowStyle="None"
-        AllowsTransparency="True"
-        Background="#80000000"
-        Topmost="True"
-        ShowInTaskbar="False"
-        WindowState="Maximized"
-        ResizeMode="NoResize"
-        Title="Playnite Overlay">
-    <Grid>
-        <Border CornerRadius="8" Background="#CC1E1E1E" Padding="24" HorizontalAlignment="Center" VerticalAlignment="Center">
-            <StackPanel Width="520" >
-                <TextBlock Text="Playnite Overlay" FontSize="24" Foreground="White" Margin="0 0 0 12"/>
-                <TextBlock Text="Use controller D-Pad/Left Stick to navigate. A = Select, B = Close" Foreground="#CCCCCC" Margin="0 0 0 16"/>
-                <UniformGrid Columns="3" Margin="0 8 0 0">
-                    <Button Content="Switch Game" Margin="6" Padding="12"/>
-                    <Button Content="Exit App" Margin="6" Padding="12"/>
-                    <Button Content="Cancel" Margin="6" Padding="12"/>
-                </UniformGrid>
-            </StackPanel>
-        </Border>
-    </Grid>
-</Window>
-
