--- src/OverlayUI/MainWindow.xaml.cs
@@ -1,12 +1,0 @@
-using System.Windows;
-
-namespace OverlayUI;
-
-public partial class MainWindow : Window
-{
-    public MainWindow()
-    {
-        InitializeComponent();
-    }
-}
-
